Create "DepositButton" component file
Related to #175
